requirements.txt
fastapi
uvicorn
python-multipart
groq
gradio
PyPDF2
python-docx
pydantic
python-dotenv